Add Vagrant file and Ansible playbook
Configure Vagrant to spin-up a VM with ubuntu 15.04 x64.
Add the new machine to the Ansible web_servers group.
Configure the Ansible playbook to add the webserver role to all machines in the web_servers group.
The webserver role should depend on the oracle-java8 role (which will automatically install the Oracle Java 8 JRE on the machine).